Changes to Booking, added new permission, fixes
Booking now allows users to bypass constraints on booking limits with a prompt.
Code for toasts has now been moved into the addBooking ajax code.
The red time line has been cleaned up so it doesn't show when it's not supposed to.
The check for conflicting bookings is much more thorough.
Resource and blocked permissions are now separate.